Revert "Update TestClass01.cs"
This reverts commit 7e1c31f1bdef53552bc759f37296fb8fb18dfe8a.

+++ TestClass01.cs
@@ -2,7 +2,7 @@
 {
     public class TestClass01
     {
-        public long Counter { get; set; }
+        public int Counter { get; set; }
         public string Name { get; set; }
     }
 }
